+++ lib/rufo/formatter.rb
@@ -2309,11 +2309,14 @@ class Rufo::Formatter
   def visit_hash(node)
     # [:hash, elements]
     _, elements = node
-
     token_column = current_token_column
+
+    closing_brace_token, _ = find_closing_brace_token
+    need_space = need_space_for_hash?(node, closing_brace_token)
 
     check :on_lbrace
     write "{"
+    write " " if need_space
     next_token
 
     if elements
@@ -2326,6 +2329,7 @@ class Rufo::Formatter
     end
 
     check :on_rbrace
+    write " " if need_space
     write "}"
     next_token
   end
@@ -3858,5 +3862,36 @@ class Rufo::Formatter
 
   def result
     @output
+  end
+
+  # Check to see if need to add space inside hash literal braces.
+  def need_space_for_hash?(node, closing_brace_token)
+    return false unless node[1]
+
+    left_need_space = current_token_line == node_line(node, beginning: true)
+    right_need_space = closing_brace_token[0][0] == node_line(node, beginning: false)
+
+    left_need_space && right_need_space
+  end
+
+  def node_line(node, beginning: true)
+    # get line of node, it is only used in visit_hash right now,
+    # so handling the following node types is enough.
+    case node.first
+    when :hash, :string_literal, :symbol_literal, :symbol, :vcall, :string_content, :assoc_splat, :var_ref
+      node_line(node[1], beginning: beginning)
+    when :assoc_new
+      node_line(beginning ? node[1] : node.last, beginning: beginning)
+    when :assoclist_from_args
+      node_line(beginning ? node[1][0] : node[1].last, beginning: beginning)
+    when :dyna_symbol
+      if node[1][0].is_a?(Symbol)
+        node_line(node[1], beginning: beginning)
+      else
+        node_line(node[1][0], beginning: beginning)
+      end
+    when :@label, :@int, :@ident, :@tstring_content, :@kw
+      node[2][0]
+    end
   end
 end